Vector pt1

+++ test/sql/aht_vector_iterator_test.cpp
@@ -22,7 +22,7 @@ namespace tpl::sql::test {
  * An input tuple, this is what we use to probe and update aggregates
  */
 struct InputTuple {
-  u64 key, col_a;
+  i64 key, col_a;
 
   InputTuple(u64 key, u64 col_a) : key(key), col_a(col_a) {}
 
@@ -35,7 +35,7 @@ struct InputTuple {
  * SELECT key, SUM(col_a), SUM(col_a*2), SUM(col_a*10) ...
  */
 struct AggTuple {
-  u64 key, count1, count2, count3;
+  i64 key, count1, count2, count3;
 
   explicit AggTuple(const InputTuple &input)
       : key(input.key), count1(0), count2(0), count3(0) {
@@ -62,10 +62,10 @@ static void Transpose(const byte **raw_aggregates, const u64 size,
   auto **aggs = reinterpret_cast<const AggTuple **>(raw_aggregates);
   for (u32 i = 0; i < size; i++, vpi->Advance()) {
     const auto *agg = aggs[i];
-    vpi->SetValue<u64, false>(0, agg->key, false);
-    vpi->SetValue<u64, false>(1, agg->count1, false);
-    vpi->SetValue<u64, false>(2, agg->count2, false);
-    vpi->SetValue<u64, false>(3, agg->count3, false);
+    vpi->SetValue<i64, false>(0, agg->key, false);
+    vpi->SetValue<i64, false>(1, agg->count1, false);
+    vpi->SetValue<i64, false>(2, agg->count2, false);
+    vpi->SetValue<i64, false>(3, agg->count3, false);
   }
 }
 
@@ -142,10 +142,10 @@ TEST_F(AggregationHashTableVectorIteratorTest, IterateSmallAggregation) {
     EXPECT_FALSE(vpi->IsFiltered());
 
     for (; vpi->HasNext(); vpi->Advance()) {
-      auto agg_key = *vpi->GetValue<u64, false>(0, nullptr);
-      auto agg_count_1 = *vpi->GetValue<u64, false>(1, nullptr);
-      auto agg_count_2 = *vpi->GetValue<u64, false>(2, nullptr);
-      auto agg_count_3 = *vpi->GetValue<u64, false>(3, nullptr);
+      auto agg_key = *vpi->GetValue<i64, false>(0, nullptr);
+      auto agg_count_1 = *vpi->GetValue<i64, false>(1, nullptr);
+      auto agg_count_2 = *vpi->GetValue<i64, false>(2, nullptr);
+      auto agg_count_3 = *vpi->GetValue<i64, false>(3, nullptr);
       EXPECT_TRUE(agg_key < num_aggs);
       EXPECT_EQ(group_size, agg_count_1);
       EXPECT_EQ(agg_count_1 * 2u, agg_count_2);
@@ -168,16 +168,16 @@ TEST_F(AggregationHashTableVectorIteratorTest, FilterPostAggregation) {
 
   PopulateAggHT(&agg_ht, num_aggs, num_tuples, 1 /* cola */);
 
-  constexpr u32 agg_needle_key = 686;
-  constexpr u32 agg_max_key = 2600;
+  constexpr i32 agg_needle_key = 686;
+  constexpr i32 agg_max_key = 2600;
 
   // Iterate
-  u32 num_needle_keys = 0, num_keys_lt_max = 0;
+  i32 num_needle_keys = 0, num_keys_lt_max = 0;
   AHTVectorIterator iter(agg_ht, output_schema(), Transpose);
   for (; iter.HasNext(); iter.Next(Transpose)) {
     auto *vpi = iter.GetVectorProjectionIterator();
     vpi->ForEach([&]() {
-      auto agg_key = *vpi->GetValue<u64, false>(0, nullptr);
+      auto agg_key = *vpi->GetValue<i64, false>(0, nullptr);
       num_needle_keys += static_cast<u32>(agg_key == agg_needle_key);
       num_keys_lt_max += static_cast<u32>(agg_key < agg_max_key);
     });
@@ -185,7 +185,7 @@ TEST_F(AggregationHashTableVectorIteratorTest, FilterPostAggregation) {
 
   // After filter, there should be exactly one key equal to the needle. Since we
   // use a dense aggregate key range
-  EXPECT_EQ(1u, num_needle_keys);
+  EXPECT_EQ(1, num_needle_keys);
   EXPECT_EQ(agg_max_key, num_keys_lt_max);
 }
 
@@ -199,7 +199,7 @@ TEST_F(AggregationHashTableVectorIteratorTest, DISABLED_Perf) {
     // Populate
     PopulateAggHT(&agg_ht, size, size * 10, 1 /* cola */);
 
-    constexpr u32 filter = 1000;
+    constexpr i32 filter = 1000;
 
     auto vaat_ms = Bench(4, [&]() {
       vaat_ret = 0;
@@ -217,7 +217,7 @@ TEST_F(AggregationHashTableVectorIteratorTest, DISABLED_Perf) {
       for (; iter.HasNext(); iter.Next()) {
         auto *agg_row =
             reinterpret_cast<const AggTuple *>(iter.GetCurrentAggregateRow());
-        if (agg_row->key < static_cast<u64>(filter)) {
+        if (agg_row->key < filter) {
           taat_ret++;
         }
       }
